refactor: merge the swc plugin
