+++ .changeset/ninety-maps-sit.md
@@ -1,0 +1,8 @@
+---
+'@modern-js/app-tools': minor
+'@modern-js/builder-plugin-swc': minor
+'@modern-js/plugin-swc': minor
+---
+
+chore: merge the swc plugin
+chore: 合并 swc 插件
