--- packages/builder/plugin-swc-base/CHANGELOG.md
@@ -1,275 +1,0 @@
-# @modern-js/builder-plugin-swc-base
-
-## 2.22.1
-
-### Patch Changes
-
-- cfcf003: fix(plugin-swc): styledComponents configuration not work
-
-  fix(plugin-swc): 修复 styledComponents 配置不生效的问题
-
-- Updated dependencies [e2848a2]
-- Updated dependencies [a470c04]
-- Updated dependencies [95ffa6b]
-- Updated dependencies [12d54ae]
-- Updated dependencies [bd4b150]
-- Updated dependencies [dd36311]
-- Updated dependencies [d4045ed]
-- Updated dependencies [8bd9981]
-- Updated dependencies [1f02cd2]
-- Updated dependencies [cfcf003]
-- Updated dependencies [15181be]
-  - @modern-js/utils@2.22.1
-  - @modern-js/builder-shared@2.22.1
-
-## 2.22.0
-
-### Patch Changes
-
-- e7a5f94: feat(plugin-swc): should reduce lodash bundle size by default
-
-  feat(plugin-swc): 默认优化 lodash 包体积
-
-- Updated dependencies [3d48836]
-- Updated dependencies [5050e8e]
-- Updated dependencies [850cde6]
-  - @modern-js/builder-shared@2.22.0
-  - @modern-js/utils@2.22.0
-
-## 2.21.1
-
-### Patch Changes
-
-- Updated dependencies [19b3b64]
-  - @modern-js/builder-shared@2.21.1
-  - @modern-js/utils@2.21.1
-
-## 2.21.0
-
-### Patch Changes
-
-- 26dcf3a: chore: bump typescript to v5 in devDependencies
-
-  chore: 升级 devDependencies 中的 typescript 版本到 v5
-
-- ad78387: chore(deps): bump babel-related dependencies to latest version
-
-  chore(deps): 升级 babel 相关依赖到最新版本
-
-- Updated dependencies [7d2972e]
-- Updated dependencies [863f8df]
-- Updated dependencies [e81eeaf]
-- Updated dependencies [26dcf3a]
-- Updated dependencies [1ef03dc]
-- Updated dependencies [056627f]
-- Updated dependencies [4274510]
-- Updated dependencies [0fc15ca]
-- Updated dependencies [43b4e83]
-- Updated dependencies [ad78387]
-  - @modern-js/builder-shared@2.21.0
-  - @modern-js/utils@2.21.0
-
-## 2.20.0
-
-### Patch Changes
-
-- Updated dependencies [3c4e0a5]
-- Updated dependencies [6b9d90a]
-  - @modern-js/builder-shared@2.20.0
-  - @modern-js/utils@2.20.0
-
-## 2.19.1
-
-### Patch Changes
-
-- Updated dependencies [afb735f]
-  - @modern-js/builder-shared@2.19.1
-  - @modern-js/utils@2.19.1
-
-## 2.19.0
-
-### Patch Changes
-
-- Updated dependencies [1134fe2]
-  - @modern-js/builder-shared@2.19.0
-  - @modern-js/utils@2.19.0
-
-## 2.18.1
-
-### Patch Changes
-
-- Updated dependencies [9b0b7ef]
-  - @modern-js/builder-shared@2.18.1
-  - @modern-js/utils@2.18.1
-
-## 2.18.0
-
-### Patch Changes
-
-- @modern-js/builder-shared@2.18.0
-- @modern-js/utils@2.18.0
-
-## 2.17.1
-
-### Patch Changes
-
-- @modern-js/builder-shared@2.17.1
-- @modern-js/utils@2.17.1
-
-## 2.17.0
-
-### Patch Changes
-
-- @modern-js/builder-shared@2.17.0
-- @modern-js/utils@2.17.0
-
-## 2.16.0
-
-### Patch Changes
-
-- 50bc0db: feat(builder): update swc-plugins, core-js, swc-helpers
-  feat(builder): 更新 swc-plugins, core-js, swc-helpers
-- Updated dependencies [fe92de6]
-- Updated dependencies [091986a]
-- Updated dependencies [5954330]
-- Updated dependencies [7596520]
-- Updated dependencies [4e876ab]
-- Updated dependencies [e4e0e01]
-  - @modern-js/builder-shared@2.16.0
-  - @modern-js/utils@2.16.0
-
-## 2.15.0
-
-### Patch Changes
-
-- Updated dependencies [3e5e790]
-  - @modern-js/builder-shared@2.15.0
-  - @modern-js/utils@2.15.0
-
-## 2.14.0
-
-### Patch Changes
-
-- Updated dependencies [b46fbcb]
-- Updated dependencies [4779152]
-- Updated dependencies [fefd1c5]
-- Updated dependencies [8a3c693]
-- Updated dependencies [9321bef]
-- Updated dependencies [9b45c58]
-- Updated dependencies [52d0cb1]
-- Updated dependencies [60a81d0]
-- Updated dependencies [864d55e]
-- Updated dependencies [b965df2]
-- Updated dependencies [dacef96]
-- Updated dependencies [16399fd]
-  - @modern-js/builder-shared@2.14.0
-  - @modern-js/utils@2.14.0
-
-## 2.13.4
-
-### Patch Changes
-
-- @modern-js/builder-shared@2.13.4
-- @modern-js/utils@2.13.4
-
-## 2.13.3
-
-### Patch Changes
-
-- Updated dependencies [18cd03f]
-- Updated dependencies [28583e8]
-  - @modern-js/builder-shared@2.13.3
-  - @modern-js/utils@2.13.3
-
-## 2.13.2
-
-### Patch Changes
-
-- @modern-js/builder-shared@2.13.2
-- @modern-js/utils@2.13.2
-
-## 2.13.1
-
-### Patch Changes
-
-- @modern-js/builder-shared@2.13.1
-- @modern-js/utils@2.13.1
-
-## 2.13.0
-
-### Patch Changes
-
-- d69c0b9: fix: pin core-js version to 3.27 in SWC
-
-  fix: SWC 相关的 core-js 版本固定到 3.27
-
-- 5deb1fb: chore: bump core-js and swc
-
-  chore: 升级 core-js 和 swc
-
-- 67a235b: fix(builder-plugin-swc): remove default env.mode
-
-  fix(builder-plugin-swc): 删除 env.mod 的默认值
-
-- Updated dependencies [1feacdc]
-- Updated dependencies [384406c]
-- Updated dependencies [c89de05]
-- Updated dependencies [348306d]
-- Updated dependencies [384e393]
-- Updated dependencies [9c0572e]
-  - @modern-js/builder-shared@2.13.0
-  - @modern-js/utils@2.13.0
-
-## 2.12.0
-
-### Patch Changes
-
-- 98f8cb6: fix(swc): can not find object_destructuring_empty.mjs
-
-  fix(swc): 修复找不到 object_destructuring_empty.mjs 的问题
-
-- Updated dependencies [c2ca6c8]
-- Updated dependencies [6d86e34]
-- Updated dependencies [fef3394]
-  - @modern-js/utils@2.12.0
-  - @modern-js/builder-shared@2.12.0
-
-## 2.11.0
-
-### Minor Changes
-
-- c6a94c7: feat: sink builder-plugin-swc abilities to `builder-plugin-swc-base`.
-  feat: 将 builder-plugin-swc 能力下沉到 builder-plugin-swc-base 里.
-- a8c08c3: feat: 添加 `source.transformImoprt`
-
-  feat: add `source.transformImoprt`
-
-### Patch Changes
-
-- 99693f0: fix: the swc compiler can not create two different options instance && move default config to swc-base
-  fix: swc 转译器无法创建两个不同配置的实体 && 将默认配置移到 swc-base
-- Updated dependencies [adcedad]
-- Updated dependencies [cfb058f]
-- Updated dependencies [a9c6083]
-- Updated dependencies [55b07fd]
-- Updated dependencies [0bd018b]
-- Updated dependencies [8b90c79]
-- Updated dependencies [a8c08c3]
-- Updated dependencies [3aa318d]
-- Updated dependencies [5d624fd]
-- Updated dependencies [e2466a1]
-- Updated dependencies [53b0a63]
-- Updated dependencies [02bb383]
-- Updated dependencies [381a3b9]
-- Updated dependencies [7a60f10]
-- Updated dependencies [cd1040f]
-- Updated dependencies [e262a99]
-- Updated dependencies [b71cef1]
-- Updated dependencies [274b2e5]
-- Updated dependencies [b9e1c54]
-  - @modern-js/builder-shared@2.11.0
-  - @modern-js/utils@2.11.0
-
-### Major Changes
-
-Initial Release
