--- packages/builder/plugin-swc-base/package.json
@@ -1,58 +1,0 @@
-{
-  "name": "@modern-js/builder-plugin-swc-base",
-  "version": "2.22.1",
-  "description": "SWC plugin for builder in Modern.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/web-infra-dev/modern.js",
-    "directory": "packages/builder/plugin-swc-base"
-  },
-  "main": "./dist/index.js",
-  "types": "./src/index.ts",
-  "scripts": {
-    "build": "tsc",
-    "dev": "tsc -w",
-    "test": "vitest run",
-    "test:watch": "vitest dev --no-coverage",
-    "test:update": "SNAPSHOT_UPDATE=1 vitest watch",
-    "prepublishOnly": "only-allow-pnpm"
-  },
-  "exports": {
-    ".": {
-      "jsnext:source": "./src/index.ts",
-      "default": "./dist/index.js",
-      "types": "./dist/index.d.ts"
-    }
-  },
-  "engines": {
-    "node": ">=12.0.0"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
-  "devDependencies": {
-    "@modern-js/builder-webpack-provider": "workspace:*",
-    "@modern-js/utils": "workspace:*",
-    "@scripts/vitest-config": "workspace:*",
-    "@types/babel__core": "^7.20.0",
-    "source-map": "^0.7.4",
-    "typescript": "^5",
-    "webpack": "^5.82.1"
-  },
-  "dependencies": {
-    "@babel/core": "^7.21.8",
-    "@babel/preset-env": "^7.21.5",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.21.5",
-    "@modern-js/builder-shared": "workspace:*",
-    "@modern-js/utils": "workspace:*",
-    "@swc/helpers": "0.5.1",
-    "core-js": "~3.30.0"
-  },
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/",
-    "access": "public",
-    "provenance": true,
-    "types": "./dist/index.d.ts"
-  }
-}
