--- packages/builder/plugin-swc-base/src/index.ts
@@ -1,5 +1,0 @@
-export * from './apply';
-export * from './checkPolyfillPlugin';
-export * from './createLoader';
-export * from './types';
-export * from './constants';
