+++ packages/builder/plugin-swc/package.json
@@ -68,9 +68,14 @@
     "react-dom": "^18"
   },
   "dependencies": {
-    "@modern-js/builder-plugin-swc-base": "workspace:*",
+    "@babel/core": "^7.21.8",
+    "@babel/preset-env": "^7.21.5",
+    "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.21.5",
+    "@swc/helpers": "0.5.1",
+    "core-js": "~3.30.0",
     "@modern-js/builder-shared": "workspace:*",
-    "@modern-js/swc-plugins": "0.3.5",
+    "@modern-js/swc-plugins": "0.4.0",
     "@modern-js/utils": "workspace:*"
   },
   "publishConfig": {
