+++ packages/builder/plugin-swc/src/checkPolyfillPlugin.ts
