+++ packages/builder/plugin-swc/src/index.ts
@@ -1,7 +1,7 @@
 export { default as SwcLoader } from './loader';
 export { builderPluginSwc, PluginSwc } from './plugin';
 export * from './binding';
-export * from './config';
+export * from './types';
 
 /// NOTE:
 /// Polyfill async function or generator, should install regenerator-runtime manually
